Python_part/requirements.txt
crewai
crewai-tools
langchain-community
langchain-core
langchain-groq
requests
python-dotenv
flask
flask-cors
reportlab